feat: new Excel upload form and API (#28105)
