+++ superset-frontend/src/views/CRUD/utils.tsx
@@ -494,7 +494,7 @@ export const uploadUserPerms = (
     findPermission('can_this_form_get', 'ColumnarToDatabaseView', roles);
   const canUploadExcel =
     checkUploadExtensions(excelExt, allowedExt) &&
-    findPermission('can_this_form_get', 'ExcelToDatabaseView', roles);
+    findPermission('can_excel_upload', 'Database', roles);
   return {
     canUploadCSV,
     canUploadColumnar,
