--- superset/templates/superset/form_view/csv_scripts.html
@@ -1,39 +1,0 @@
-{#
-Licensed to the Apache Software Foundation (ASF) under one
-or more contributor license agreements. See the NOTICE file
-distributed with this work for additional information
-regarding copyright ownership. The ASF licenses this file
-to you under the Apache License, Version 2.0 (the
-"License"); you may not use this file except in compliance
-with the License. You may obtain a copy of the License at
-
-http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing,
-software distributed under the License is distributed on an
-"AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-KIND, either express or implied. See the License for the
-specific language governing permissions and limitations
-under the License.
-#}
-{% import "superset/macros.html" as macros %}
-
-<script nonce="{{ macros.get_nonce() }}">
-  $('#delimiter').on('change', function () {
-    var delimiterOptions = $('#delimiter').val();
-    if (delimiterOptions?.includes("other")) {
-      document.getElementById("otherInput").style.display = 'block';
-      $('#otherInput').attr('required', 'required');
-    } else {
-      document.getElementById("otherInput").style.display = 'none';
-      $('#otherInput').removeAttr('required');
-    }
-  }).change();
-
-  $(".collapse").on("hide.bs.collapse show.bs.collapse", e => {
-    $(e.target)
-      .prev()
-      .find("i:last-child")
-      .toggleClass("fa-chevron-up fa-chevron-down");
-  });
-</script>
